Removing a bunch of unused rails helpers, views etc
