--- server/src/main/webapp/WEB-INF/rails/app/models/config_update/save_action.rb
@@ -1,29 +1,0 @@
-#
-# Copyright 2019 ThoughtWorks, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-module ConfigUpdate
-  class SaveAction
-    include com.thoughtworks.go.config.update.UpdateConfigFromUI
-    include ::ConfigUpdate::RefsAsUpdatedRefs
-    attr_reader :params
-
-    def initialize params, user, security_service
-      @params = params
-      @user = user
-      @security_service = security_service
-    end
-  end
-end
