--- server/src/main/webapp/WEB-INF/rails/app/models/package_view_model.rb
@@ -1,37 +1,0 @@
-#
-# Copyright 2019 ThoughtWorks, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-class PackageViewModel
-  attr_accessor :properties, :name
-
-  def initialize(metadata, package)
-    @properties = []
-    if(package)
-      @name = package.getName()
-      return if metadata.nil?
-      package_configurations = metadata.list()
-      package_configurations.each do |config|
-        property = package.getConfiguration().getProperty(config.getKey())
-        @properties << PackagePropertyModel.new(config, property)
-      end
-    end
-  end
-
-  def filterSecureProperties!
-    @properties.reject! { |x| x.is_secure }
-    self
-  end
-end
