--- server/src/main/webapp/WEB-INF/rails/app/views/admin/shared/_form_submit.html.erb
@@ -1,4 +1,0 @@
-<div class="form_buttons">
-    <%= submit_button('SAVE', :class => 'primary skip_dirty_stop finish') %>
-    <%= link_to 'RESET', scope[:reset_url], :class => 'reset_button link_as_button skip_dirty_stop', :id => 'reset_form' -%>
-</div>
