--- server/src/main/webapp/WEB-INF/rails/app/views/analytics/index.html.erb
@@ -1,1 +1,0 @@
-<div class="analytics-container" data-pipeline-list="<%= @pipeline_list.to_json %>" />
