--- server/src/main/webapp/WEB-INF/rails/app/views/api/dependency_modifications/index.xml.erb
@@ -1,11 +1,0 @@
-<modifications>
-    <title><%= cdata_section(params[:pipeline_name] + "/" + params[:stage_name]) -%></title>
-
-    <%- @modifications.each do |mod| -%>
-        <entry>
-            <revision><%= cdata_section(mod.getRevision()) -%></revision>
-            <modifiedTime><%= mod.getModifiedTime().iso8601 -%></modifiedTime>
-            <pipelineLabel><%= cdata_section(mod.getPipelineLabel()) -%></pipelineLabel>
-        </entry>
-    <%- end -%>
-</modifications>
