--- server/src/main/webapp/WEB-INF/rails/app/views/api/users/index.xml.erb
@@ -1,1 +1,0 @@
-<%== pretty_print_xml(@doc) -%>
