--- server/src/main/webapp/WEB-INF/rails/spec/lib/config_update/save_action_spec.rb
@@ -1,35 +1,0 @@
-#
-# Copyright 2019 ThoughtWorks, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-require 'rails_helper'
-
-describe ConfigUpdate::SaveAction do
-
-  it "should create a valid object" do
-    params = {:a => :b}
-    security_service = Object.new
-    save_action = ::ConfigUpdate::SaveAction.new(params, "loser", security_service)
-
-    expect(save_action.params).to eq(params)
-    expect(save_action.instance_variable_get("@user")).to eq("loser")
-    expect(save_action.instance_variable_get("@security_service")).to eq(security_service)
-  end
-
-  it "should create ConfigUpdate command" do
-    save_action = ::ConfigUpdate::SaveAction.new(nil, nil, nil)
-    expect(save_action.java_kind_of?(com.thoughtworks.go.config.update.UpdateConfigFromUI)).to be_truthy
-  end
-end
