--- server/src/main/webapp/WEB-INF/rails/spec/lib/config_update/save_as_super_admin_spec.rb
@@ -1,25 +1,0 @@
-#
-# Copyright 2019 ThoughtWorks, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-require 'rails_helper'
-
-describe ConfigUpdate::SaveAsSuperAdmin do
-
-  it "should check if user is super amdin" do
-    save = ::ConfigUpdate::SaveAsSuperAdmin.new(nil, nil, nil)
-    expect(save.class.included_modules).to include(::ConfigUpdate::CheckIsSuperAdmin)
-  end
-end
